Move timer* files to submodule, split out common code
