+++ esp-wifi/src/lib.rs
@@ -51,12 +51,6 @@ mod binary {
 mod compat;
 mod preempt;
 
-#[cfg_attr(esp32, path = "timer_esp32.rs")]
-#[cfg_attr(esp32c3, path = "timer_esp32c3.rs")]
-#[cfg_attr(esp32c2, path = "timer_esp32c2.rs")]
-#[cfg_attr(esp32c6, path = "timer_esp32c6.rs")]
-#[cfg_attr(esp32s3, path = "timer_esp32s3.rs")]
-#[cfg_attr(esp32s2, path = "timer_esp32s2.rs")]
 mod timer;
 
 #[cfg(feature = "wifi")]
