+++ esp-wifi/src/timer/timer_esp32c2.rs
@@ -1,7 +1,3 @@
-use core::cell::RefCell;
-
-use critical_section::Mutex;
-
 use crate::{
     binary,
     hal::{
@@ -11,29 +7,13 @@ use crate::{
         riscv,
         systimer::{Alarm, Periodic, SystemTimer, Target},
     },
-    preempt::preempt::task_switch,
 };
 
-pub const TICKS_PER_SECOND: u64 = 16_000_000;
-
-pub const COUNTER_BIT_MASK: u64 = 0x000F_FFFF_FFFF_FFFF;
-
-const TIMER_DELAY: fugit::HertzU32 = fugit::HertzU32::from_raw(crate::CONFIG.tick_rate_hz);
-
-static ALARM0: Mutex<RefCell<Option<Alarm<Periodic, 0>>>> = Mutex::new(RefCell::new(None));
+pub use super::arch_specific::{get_systimer_count, yield_task, TICKS_PER_SECOND};
+use super::arch_specific::{setup_multitasking, setup_timer};
 
 pub fn setup_timer_isr(systimer: Alarm<Target, 0>) {
-    let alarm0 = systimer.into_periodic();
-    alarm0.set_period(TIMER_DELAY.into());
-    alarm0.clear_interrupt();
-    alarm0.interrupt_enable(true);
-
-    critical_section::with(|cs| ALARM0.borrow_ref_mut(cs).replace(alarm0));
-
-    unwrap!(interrupt::enable(
-        Interrupt::SYSTIMER_TARGET0,
-        interrupt::Priority::Priority1,
-    ));
+    setup_timer(systimer);
 
     #[cfg(feature = "wifi")]
     {
@@ -59,16 +39,7 @@ pub fn setup_timer_isr(systimer: Alarm<Target, 0>) {
         ));
     }
 
-    unwrap!(interrupt::enable(
-        Interrupt::FROM_CPU_INTR3,
-        interrupt::Priority::Priority1,
-    ));
-
-    unsafe {
-        riscv::interrupt::enable();
-    }
-
-    while unsafe { crate::preempt::FIRST_SWITCH.load(core::sync::atomic::Ordering::Relaxed) } {}
+    setup_multitasking();
 }
 
 #[cfg(feature = "wifi")]
@@ -147,48 +118,4 @@ fn BT_MAC() {
 
         trace!("interrupt BT_MAC done");
     };
-}
-
-#[interrupt]
-fn SYSTIMER_TARGET0(trap_frame: &mut TrapFrame) {
-    // clear the systimer intr
-    critical_section::with(|cs| {
-        unwrap!(ALARM0.borrow_ref_mut(cs).as_mut()).clear_interrupt();
-    });
-
-    task_switch(trap_frame);
-}
-
-#[interrupt]
-fn FROM_CPU_INTR3(trap_frame: &mut TrapFrame) {
-    unsafe {
-        // clear ETS_FROM_CPU_INTR3
-        (&*peripherals::SYSTEM::PTR)
-            .cpu_intr_from_cpu_3
-            .modify(|_, w| w.cpu_intr_from_cpu_3().clear_bit());
-    }
-
-    critical_section::with(|cs| {
-        let mut alarm0 = ALARM0.borrow_ref_mut(cs);
-        let alarm0 = unwrap!(alarm0.as_mut());
-
-        alarm0.set_period(TIMER_DELAY.into());
-        alarm0.clear_interrupt();
-    });
-
-    task_switch(trap_frame);
-}
-
-pub fn yield_task() {
-    unsafe {
-        (&*peripherals::SYSTEM::PTR)
-            .cpu_intr_from_cpu_3
-            .modify(|_, w| w.cpu_intr_from_cpu_3().set_bit());
-    }
-}
-
-/// Current systimer count value
-/// A tick is 1 / 16_000_000 seconds
-pub fn get_systimer_count() -> u64 {
-    SystemTimer::now()
 }
