+++ esp-wifi/src/timer/xtensa.rs
@@ -1,27 +1,27 @@
 use core::cell::RefCell;
 
-use crate::hal::{
-    interrupt,
-    macros::interrupt,
-    peripherals::{self, TIMG1},
+use crate::{
+    hal::{
+        interrupt,
+        macros::interrupt,
+        peripherals::{self, TIMG1},
+        prelude::*,
+        timer::{Timer, Timer0},
+        trapframe::TrapFrame,
+        xtensa_lx, xtensa_lx_rt,
+    },
     preempt::preempt::task_switch,
-    prelude::*,
-    timer::{Timer, Timer0},
-    trapframe::TrapFrame,
-    xtensa_lx, xtensa_lx_rt,
 };
 use atomic_polyfill::{AtomicU32, Ordering};
 use critical_section::Mutex;
 
-pub const TICKS_PER_SECOND: u64 = 40_000_000;
-
-pub const COUNTER_BIT_MASK: u64 = 0xFFFF_FFFF_FFFF_FFFF;
-
-const TIMER_DELAY: fugit::HertzU32 = fugit::HertzU32::from_raw(crate::CONFIG.tick_rate_hz);
-
 static TIMER1: Mutex<RefCell<Option<Timer<Timer0<TIMG1>>>>> = Mutex::new(RefCell::new(None));
 
 static TIMER_OVERFLOWS: AtomicU32 = AtomicU32::new(0);
+
+const TIMER_DELAY: fugit::HertzU32 = fugit::HertzU32::from_raw(crate::CONFIG.tick_rate_hz);
+
+pub const TICKS_PER_SECOND: u64 = 40_000_000;
 
 /// This function must not be called in a critical section. Doing so may return an incorrect value.
 pub fn get_systimer_count() -> u64 {
@@ -41,36 +41,13 @@ pub fn get_systimer_count() -> u64 {
     (((overflow as u64) << 32) + counter_after as u64) * 40_000_000 / 240_000_000
 }
 
-pub fn setup_timer_isr(timg1_timer0: Timer<Timer0<TIMG1>>) {
+pub fn setup_timer(timg1_timer0: Timer<Timer0<TIMG1>>) {
     let mut timer1 = timg1_timer0;
+
     unwrap!(interrupt::enable(
         peripherals::Interrupt::TG1_T0_LEVEL,
         interrupt::Priority::Priority2,
     ));
-
-    #[cfg(feature = "wifi")]
-    {
-        unwrap!(interrupt::enable(
-            peripherals::Interrupt::WIFI_MAC,
-            interrupt::Priority::Priority1,
-        ));
-        unwrap!(interrupt::enable(
-            peripherals::Interrupt::WIFI_PWR,
-            interrupt::Priority::Priority1,
-        ));
-    }
-
-    #[cfg(feature = "ble")]
-    {
-        unwrap!(interrupt::enable(
-            peripherals::Interrupt::BT_BB,
-            interrupt::Priority::Priority1,
-        ));
-        unwrap!(interrupt::enable(
-            peripherals::Interrupt::RWBLE,
-            interrupt::Priority::Priority1,
-        ));
-    }
 
     timer1.listen();
     timer1.start(TIMER_DELAY.into_duration());
@@ -79,7 +56,9 @@ pub fn setup_timer_isr(timg1_timer0: Timer<Timer0<TIMG1>>) {
     });
 
     xtensa_lx::timer::set_ccompare0(0xffffffff);
+}
 
+pub fn setup_multitasking() {
     unsafe {
         let enabled = xtensa_lx::interrupt::disable();
         xtensa_lx::interrupt::enable_mask(
@@ -99,63 +78,6 @@ fn Timer0(_level: u32) {
     TIMER_OVERFLOWS.fetch_add(1, Ordering::Relaxed);
 
     xtensa_lx::timer::set_ccompare0(0xffffffff);
-}
-
-#[cfg(feature = "wifi")]
-#[interrupt]
-fn WIFI_MAC() {
-    unsafe {
-        let (fnc, arg) = crate::wifi::os_adapter::ISR_INTERRUPT_1;
-        trace!("interrupt WIFI_MAC {:?} {:?}", fnc, arg);
-
-        if !fnc.is_null() {
-            let fnc: fn(*mut crate::binary::c_types::c_void) = core::mem::transmute(fnc);
-            fnc(arg);
-        }
-    }
-}
-
-#[cfg(feature = "wifi")]
-#[interrupt]
-fn WIFI_PWR() {
-    unsafe {
-        let (fnc, arg) = crate::wifi::os_adapter::ISR_INTERRUPT_1;
-        trace!("interrupt WIFI_PWR {:?} {:?}", fnc, arg);
-
-        if !fnc.is_null() {
-            let fnc: fn(*mut crate::binary::c_types::c_void) = core::mem::transmute(fnc);
-            fnc(arg);
-        }
-
-        trace!("interrupt 1 done");
-    };
-}
-
-#[cfg(feature = "ble")]
-#[interrupt]
-fn RWBLE() {
-    critical_section::with(|_| unsafe {
-        let (fnc, arg) = crate::ble::btdm::ble_os_adapter_chip_specific::ISR_INTERRUPT_5;
-        trace!("interrupt RWBLE {:?} {:?}", fnc, arg);
-        if !fnc.is_null() {
-            let fnc: fn(*mut crate::binary::c_types::c_void) = core::mem::transmute(fnc);
-            fnc(arg);
-        }
-    });
-}
-
-#[cfg(feature = "ble")]
-#[interrupt]
-fn BT_BB() {
-    critical_section::with(|_| unsafe {
-        let (fnc, arg) = crate::ble::btdm::ble_os_adapter_chip_specific::ISR_INTERRUPT_8;
-        trace!("interrupt RWBT {:?} {:?}", fnc, arg);
-
-        if !fnc.is_null() {
-            let fnc: fn(*mut crate::binary::c_types::c_void) = core::mem::transmute(fnc);
-            fnc(arg);
-        }
-    });
 }
 
 #[interrupt]
